+++ app/src/main/kotlin/com/yourssu/soomsil/usaint/data/repository/StudentInfoRepository.kt
@@ -1,9 +1,9 @@
 package com.yourssu.soomsil.usaint.data.repository
 
-import com.yourssu.soomsil.usaint.data.model.StudentInfoDto
 import com.yourssu.soomsil.usaint.data.source.local.datastore.StudentInfoDataStore
 import com.yourssu.soomsil.usaint.data.source.remote.rusaint.RusaintApi
-import com.yourssu.soomsil.usaint.domain.type.UserCredential
+import com.yourssu.soomsil.usaint.model.StudentInfoDto
+import com.yourssu.soomsil.usaint.model.UserCredential
 import dev.eatsteak.rusaint.ffi.USaintSession
 import timber.log.Timber
 import javax.inject.Inject
