+++ app/src/main/kotlin/com/yourssu/soomsil/usaint/domain/usecase/LecturesDiffUseCase.kt
@@ -1,9 +1,24 @@
 package com.yourssu.soomsil.usaint.domain.usecase
 
 import com.yourssu.soomsil.usaint.data.source.local.entity.LectureVO
-import com.yourssu.soomsil.usaint.domain.type.LectureDiff
-import com.yourssu.soomsil.usaint.domain.type.diff
 import javax.inject.Inject
+
+// Pair<before, after>
+data class LectureDiff(
+    val title: String,
+    val code: String,
+    val credit: Pair<Float, Float>?,
+    val grade: Pair<String, String>?,
+    val score: Pair<String, String>?,
+)
+
+fun LectureVO.diff(other: LectureVO) = LectureDiff(
+    title = title,
+    code = code,
+    credit = if (credit == other.credit) null else (credit to other.credit),
+    grade = if (grade == other.grade) null else (grade to other.grade),
+    score = if (score == other.score) null else (score to other.score),
+)
 
 class LecturesDiffUseCase @Inject constructor() {
     operator fun invoke(old: List<LectureVO>, new: List<LectureVO>): List<LectureDiff> {
