+++ app/src/main/kotlin/com/yourssu/soomsil/usaint/screen/home/HomeViewModel.kt
@@ -9,12 +9,12 @@ import com.yourssu.soomsil.usaint.data.repository.StudentInfoRepository
 import com.yourssu.soomsil.usaint.data.repository.TotalReportCardRepository
 import com.yourssu.soomsil.usaint.data.repository.USaintSessionRepository
 import com.yourssu.soomsil.usaint.screen.UiEvent
-import com.yourssu.soomsil.usaint.ui.entities.ReportCardSummary
-import com.yourssu.soomsil.usaint.ui.entities.StudentInfo
-import com.yourssu.soomsil.usaint.ui.entities.toCredit
-import com.yourssu.soomsil.usaint.ui.entities.toGrade
-import com.yourssu.soomsil.usaint.ui.entities.toReportCardSummary
-import com.yourssu.soomsil.usaint.ui.entities.toStudentInfo
+import com.yourssu.soomsil.usaint.ui.types.ReportCardSummary
+import com.yourssu.soomsil.usaint.ui.types.StudentInfo
+import com.yourssu.soomsil.usaint.ui.types.toCredit
+import com.yourssu.soomsil.usaint.ui.types.toGrade
+import com.yourssu.soomsil.usaint.ui.types.toReportCardSummary
+import com.yourssu.soomsil.usaint.ui.types.toStudentInfo
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.eatsteak.rusaint.ffi.RusaintException
 import kotlinx.coroutines.Job
