+++ app/src/main/kotlin/com/yourssu/soomsil/usaint/screen/semesterdetail/SemesterDetailViewModel.kt
@@ -13,11 +13,11 @@ import com.yourssu.soomsil.usaint.domain.type.SemesterType
 import com.yourssu.soomsil.usaint.domain.usecase.GetCurrentSemesterTypeUseCase
 import com.yourssu.soomsil.usaint.domain.usecase.MakeSemesterFromLecturesUseCase
 import com.yourssu.soomsil.usaint.screen.UiEvent
-import com.yourssu.soomsil.usaint.ui.entities.LectureInfo
-import com.yourssu.soomsil.usaint.ui.entities.Semester
-import com.yourssu.soomsil.usaint.ui.entities.sortByGrade
-import com.yourssu.soomsil.usaint.ui.entities.toLectureInfo
-import com.yourssu.soomsil.usaint.ui.entities.toSemester
+import com.yourssu.soomsil.usaint.ui.types.LectureInfo
+import com.yourssu.soomsil.usaint.ui.types.Semester
+import com.yourssu.soomsil.usaint.ui.types.sortByGrade
+import com.yourssu.soomsil.usaint.ui.types.toLectureInfo
+import com.yourssu.soomsil.usaint.ui.types.toSemester
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.eatsteak.rusaint.ffi.USaintSession
 import kotlinx.coroutines.flow.MutableSharedFlow
