+++ app/src/main/kotlin/com/yourssu/soomsil/usaint/screen/semesterdetail/components/GradeSummary.kt
@@ -20,12 +20,12 @@ import androidx.compose.ui.unit.dp
 import com.yourssu.soomsil.usaint.R
 import com.yourssu.soomsil.usaint.domain.type.makeSemesterType
 import com.yourssu.soomsil.usaint.screen.semesterdetail.CaptureFlag
-import com.yourssu.soomsil.usaint.ui.entities.Credit
-import com.yourssu.soomsil.usaint.ui.entities.Grade
-import com.yourssu.soomsil.usaint.ui.entities.Semester
-import com.yourssu.soomsil.usaint.ui.entities.toCredit
-import com.yourssu.soomsil.usaint.ui.entities.toGrade
 import com.yourssu.soomsil.usaint.ui.theme.SoomsilUSaintTheme
+import com.yourssu.soomsil.usaint.ui.types.Credit
+import com.yourssu.soomsil.usaint.ui.types.Grade
+import com.yourssu.soomsil.usaint.ui.types.Semester
+import com.yourssu.soomsil.usaint.ui.types.toCredit
+import com.yourssu.soomsil.usaint.ui.types.toGrade
 import java.text.DecimalFormat
 
 @Composable
