--- app/src/main/kotlin/com/yourssu/soomsil/usaint/ui/component/chart/ChartData.kt
@@ -1,7 +1,0 @@
-package com.yourssu.soomsil.usaint.ui.component.chart
-
-import com.yourssu.soomsil.usaint.ui.entities.Semester
-
-data class ChartData(
-    val semesters: List<Semester>,
-)
