+++ app/src/main/kotlin/com/yourssu/soomsil/usaint/ui/types/Credit.kt
@@ -1,4 +1,4 @@
-package com.yourssu.soomsil.usaint.ui.entities
+package com.yourssu.soomsil.usaint.ui.types
 
 import androidx.compose.runtime.Immutable
 import java.text.DecimalFormat
